Move logging config options into the log module
We learned a lesson in Nova - it's best to declare and use config
options within a single module if possible. Globally declared and
use config options grow like weeds and it becomes harder to find out
if, where and how individual options are used.

Strangely, in cfg itself, we randomly declare a bunch of logging options
which are only used within the openstack.common.log module - let's move
the options there and remove the CommonConfigOpts class before they
become part of the API we commit to when oslo-config is released.

A minor detail in the patch - the logfile and logdir options are already
deprecated in favour of log_file and log_dir, but we never got around to
removing all other traces of the deprecated options.

Change-Id: I3913ea54465658d93dc56e014dfe5d911b0541d6