Gracefully handle errors in logging config files
Fixes bug #1154245

If the parameter to the --log-config option cannot be loaded by the
logging.config.fileConfig() function, you get an exception like this
raised:

  NoSectionError: No section: 'formatters'

which doesn't do much to help users understand the error.

Improve the situation by wrapping the error in a custom exception type
giving this:

 LogConfigError: Error loading logging config /etc/logging.conf: No section: 'formatters'

Also add some tests to check we raise this error under the most common
failure conditions.

Change-Id: I6ad2eb4867213b6356ce43d75da50218d1b498f6