Add error type to unhandled exception log message
Whenever an unhandled exception is thrown, only the exception value is printed
as part of the log message. Printing just the value is sometimes not very
helpful to identify the problem. eg., in case of KeyError, only the 'key_name'
would be printed as part of the log message.
This commit includes the exception type along with value while logging an
unhandled exception.

Change-Id: Ibf00173a07510b2ec0c81a29624004f17810a4e4
Closes-Bug: #1265561