default connectionpool to WARN log level
requests.packages.urllib3.connectionpool issues an INFO log message
every time a new connection is added to the pool. While this may
be incredibly interesting to the authors of the requests package,
to an OpenStack operator this is just log spam.

set the default for this logger to WARN to tighten up the logs so
that INFO level is meaningful.

Change-Id: Icb30afe569ba80d1d322ef7f22a46a3e60b3ba9a