always log a traceback in the sys.excepthook
oslo logging overwrites the system sys.excepthook. However, in
the default case it does so in a way that masks the actual exception.
This means that developing with oslo logging enabled is really
really frustrating, as python syntax errors are just randomly
swallowed and you are off to guessing.

Because it's taking over the system level excepthook, it needs to
not do the very unexpected thing and eat all the real information.

Change-Id: I8d03d4abd77453856191c48512b7a27a7773084b