log: make set_defaults() tests clean up properly
A call to log.set_defaults() changes the 'default' attribute on the
option definitions statically defined in the log module. Currently, the
tests for this function do not restore the original default values when
the test completes.

This only becomes an issue when the tests are called in a sequence like
this:

  - test_change_default_log_level() sets default_log_levels=['foo=bar']

  - test_error_notification() subsequently calls log.setup() which tries
    to do logger.setLevel('bar') because it's using the bogus default
    from the previous test

Change-Id: Id020483049f46b78f834381f37d4ad6c75d49d25