Initial translation setup
To start translation, we need to initially import the
translation file - and place it at the proper place so that
the usual CI scripts can handle it.

The proper place is for all python projects
$PROJECT/locale/$PROJECT.pot - see setup.cfg.

Further imports of the pot files and its translations
will be done by the OpenStack Proposal bot in a daily job.

Change-Id: Ic39a4cfe40ccdff4d6a299189fcfc0e6a9ec9c61