Switch from ContextAdapter to ContextFormatter
Remove the ContextAdapter and update getLogger to return
KeywordArgumentAdapter instances instead.

Update the JSONFormatter to do some context magic like the
ContextFormatter does so the extra properties from the context and the
KeywordArgumentAdapter are included in the output JSON package.

Update the ContextFormatter to include some nova-specific behavior that
was in ContextAdapter. We will replace that with changes to the way the
default log format strings are constructed in a later patch.

bp/remove-context-adapter

Change-Id: I5a96d55355bf4ef68f2ee71160a3e0200c371e87