Use RequestContext store in oslo_context
Remove the _local store as we don't need it anymore. We should
fetch the context directly from oslo_context's context module
using the get_current() method. Use the fixture in oslo_context
to adapt the test cases to work with the change outlined above.

Since we are using RequestContext from oslo_context, adjust a
bit of code to use resource_uuid in addition to instance_uuid
as well. Added a fixture in the base test case itself to make
sure that the context thread local store is cleared in between
tests

Change-Id: I09442d824400041768dbcddfbce932b175c767ab