Pickup instance from log format record
Nova has moved to oslo.log, however one specific usecase is
blowing up in Nova. when they use instance=instance in say
LOG.info to pass in instance, they expect the uuid in the
instance dict/object to be printed in the log. However we
end up with the string-ified version of the Instance object
or the entire instance dict which is not good.

Closes-Bug: #1426544
Change-Id: I2d1b09015bead2012327f973f45f3f6728160c69