Restore automatic unicode conversion
Earlier versions of oslo.log tried to assume that getLogger() was not
generally being used and so we relied on the developer to ensure that
their log messages were valid unicode strings. Since then we have found
a lot of cases where non-unicode (indeed, non-string) objects are passed
to loggers as messages. To avoid the resulting unicode-related errors,
this patch restores the old behavior of having the adapter convert
messages to unicode before passing them down the stack to the
logger. The caveat that comes with relying on this is that it only works
for OpenStack code that calls oslo_log.log.getLogger, and not the
standard libary's logging.getLogger.

Change-Id: Ia74f4ea043c293106bf962fb4d5bd52065e42213