Provide an API to let tempest control the log file
By default we log to stdout, but tempest wants to provide a log
file. Because tempest doesn't use a configuration file, there's no easy
way for them to give us a filename to use. We don't want most projects
to set a programmatic default, so rather than adding this to
set_defaults(), we define a new function and document that only tempest
should call it.

Change-Id: I1d4919ae1bb810931ac06226165bbf1f31e25417