Set verbose to True and deprecate it
Defaulting log level to WARNING means we store nearly no information
in logs. INFO logging contains all state transition information,
which is completely critical for understanding what is going on
in a system. Of all people I personally asked, everybody is using
DEBUG level in production, so INFO does not seem to be an overkill.

Please follow this mailing list thread for more detailed reasoning:
http://lists.openstack.org/pipermail/openstack-dev/2015-July/070609.html

Change-Id: I306535c6ca5dbdaf9398b44697578a3a30e52111