Allow reload of 'debug' option
Ie9e9a895da1a3fc62580c1febb7cd9f965bbc0f5 adds a 'mutate' flag to
options. This is the first patch to use this functionality. It marks the
'debug' option as mutable and registers '_mutate_hook' to be called
when mutate_config_files runs.

Option 'verbose' is deprecated for removal so I see no need to mark it
mutable even though it's closely related.

Change-Id: I1e7a69de169cc85f4c09954b2f46ce2da7106d90
Depends-On: Ied006631a6edbeeffae485d28eff700b13a626c1
Depends-On: I9bb2ff39dc1ed0a63bd7e7635704b34c53d32d79