Remove 'verbose' option (again)
The deprecation for this was merged on aug 1, 2015 in review:
https://review.openstack.org/#/c/206437/
That was a long time ago so it should be possible to remove it now.

This was already merged once but had to be reverted because some
projects were still relying on it. Let's try again, this time using
INFO default not WARNING, thanks dtantsur!

Depends-On: I713e9e5107a64ec29c220d1150c1bf5fcf6f1135
Depends-On: I102e8c7bca508b17ce2987b33148e99c79d0c147

Change-Id: I9e2bf358a0b7a2413388007b33902b91de522b22