Reduce READ_FREQ and TIMEOUT for watch-file
The watch-file option currently relies on pyinotify which makes
use of select.poll. This is eventlet incompatible so it blocks
the whole thread for TIMEOUT milliseconds every READ_FREQ seconds.

This meant that an agent/service using this option is blocked for
500 milliseconds every 1 second.

Until a non-blocking alternative can be found, this adjusts the
READ_FREQ and TIMEOUT params on the FastWatchedFileHandler to
3 and 5 respectively. So every 3 seconds, it will block for 5
milliseconds so only 0.16% of time is lost instead of 50%.

Change-Id: I6fdc0a3f1eaae39eb86440c6c4b07ca8d91f2462
Partial-Bug: #1583270
(cherry picked from commit 6a36cffd9fdbedfe859ad52f4e8c44897f374cbd)