Default use_stderr to False
While defaulting use_stderr to True might be a useful development
default, it is not useful for operators. Operators using systemd,
upstart, or any of these kinds of services, will find that if they
specify --log-file they will have two copies of the same logs on their
disk:

- one in the service log (e.g., /var/log/upstart)
- one in the log file they specify with --log-file

Instead of operators having to disable stderr log in every
configuration, we should just default it to False here.

Closes-bug #1588051

Change-Id: I85f8670d98d03101888fb7bb0c89eb809719b4e8