protect systemd class initialization when syslog is not available
The systemd handler needs the syslog level values for its setup, but the
library is used on platforms where syslog is not present. Ignore the
situation when setting up the class, but throw an error if the class is
used at runtime.

This also converts the exception for the missing systemd library to a
RuntimeException so the two errors for missing libraries are consistent.

Change-Id: I76729503cb2ec9df878b19a1d76786c3efd9054e
Signed-off-by: Doug Hellmann <doug@doughellmann.com>