formatter: skip ImportError when adding error_summary
In some projects, we conditionally import modules that may be not
installed, for example in Neutron where we import from ibm_db_alembic
but ignore ImportError in alembic env.py.

It turned out that ImportError is not cleared by except: pass in py2
(but is correctly cleared in py3), as can be seen in the snippet below
(kudos to @dhellmann for providing it):

    apu:~$ cat testme.py
    import sys

    print('before:', sys.exc_info()[0])

    try:
        import nonesuch
    except ImportError:
        pass

    print('after:', sys.exc_info()[0])
    apu:~$ python2 ./testme.py
    ('before:', None)
    ('after:', <type 'exceptions.ImportError'>)
    apu:~$ python3 ./testme.py
    before: None
    after: None

The patch makes the formatter skip ImportErrors. We could in theory
limit it to py2 only but to simplify matters and to stick to identical
behavior across python runtimes, we will just unconditionally skip it
for all library consumers.

This is a follow up to Ifecb831c0adf4086a9d39feb3eb7e3865db780e6 ("skip
built-in exceptions when adding error_summary") that already skipped
some builtin exceptions.

Change-Id: Ia8f7e2501a49345b8d828ec3ee4a7cc870c8a0a8