Truncate error_summary if exc_info not explicitly passed
Currently, when logging at level >= WARNING, exception information is
added to the log message as an "error summary" for operators. The error
summary includes the exception type and the message. However, when the
exception is a _Remote exception generated from oslo.messaging and sent
over RPC, the full traceback from the exception is appended to the
exception message [1] to preserve it. So, the error summary ends up
outputting the full traceback, unintentionally.

This truncates the error_summary to keep only the first line unless
exc_info was explicitly passed, via LOG.exception(), for example.

Closes-Bug: #1729475

[1] https://github.com/openstack/oslo.messaging/blob/8d42484/oslo_messaging/_drivers/common.py#L208

Change-Id: I2e4e356d23dff2e6ac1946fe9e994c4d962ab55a