Fix Formatter subclasses for Python 3.2+
Python 3.2 added the 'style' parameter to 'logging.Formatter.__init__'.
This is provided by 'logging.config.fileConfig' meaning this currently
raises a 'TypeError' similar to the below when the incompatible
formatters, JSONFormatter and FluentFormatter, are used with Python 3.2
or greater.

  TypeError: __init__() got an unexpected keyword argument 'style'

Resolve this by simply adding the parameter to the list. There is more
work we can do here (like actually supporting this parameter) but that's
a job for another patch.

Note that we can't actually test this as doing so would involve invoking
'logging.config.fileConfig', which in turn would modify the global state
of the 'logging' module. You can thank the singleton pattern for that.

Change-Id: I9b339163ddfe440bc6782ced33595a0dcf60f658
Closes-Bug: 1739743
Co-Authored-By: Stephen Finucane <stephenfin@redhat.com>