Add devstack job with JSONFormatter configured
We've run into issues in the past where a service passed something
to the logger that broke JSONFormatter. To try to catch those sooner,
add a job that configures the services to use JSONFormatter. This
should provide a more realistic test of the formatter than we can
hope to accomplish in unit tests.

Change-Id: Icfb399cfe3dce89dfd5fb5079295a4947828417a