Provide reset_color key on log record
When logging with color, it is necessary to reset the color at the
end of each line to avoid the color bleeding into subsequent log
lines.  This change adds a new key to the record called 'reset_color'
that serves as a shortcut to the escape sequence to reset the color.
It also provides documentation for the ColorHandler class that
explains how to use it.

Change-Id: I68f1c716cbed241f79fa65dae1affe810b8e6e30
Closes-Bug: 1731477