Automatically append reset_color to log lines
It's confusing that log colors can persist across multiple log lines
if they are not explicitly reset at the end of each line.  This
change automatically appends the reset escape sequence to the end of
each log record passed through the ColorHandler so the behavior is
more what most people would expect.

Change-Id: I5fefdbdd7dd2ca437863bf2117b5a93b527e9776