rewrite tests to not rely on implementation details of logging module
The tests for ensuring the results of the adapter are passed to the
logger correctly rely more on the implementation details of the
logging library than on the adapter supposedly being tested. As the
underlying implementation changes in various ways, this causes the
tests to fail (most recently when we try to run the tests under python
3.6. Rather than adding yet another case to the conditional for this
specific version of python, rewrite the tests so they do not rely on
the underlying implementation details.

Change-Id: Id4d416852553b8d11ad27e25edcf9cceab3eb1b6
Signed-off-by: Doug Hellmann <doug@doughellmann.com>

fix

Change-Id: I1f5140b0a9499c94e13fa3cd0f7151dc6c952621
Signed-off-by: Doug Hellmann <doug@doughellmann.com>