Default oslo.policy logging to INFO
As of change https://review.openstack.org/619260/ to oslo.policy
a lot of information is logged at debug for every policy check,
which gets to be a lot of information in something like the
nova-api logs which does a policy check per API request. So this
change makes the default log level INFO for oslo.policy.

Change-Id: Ie4a114219a69390efd148b705fad05fd8b9a42de
Related-Bug: #1807747