Clarify some config options
current help message to log-config-append is a bit misleading:
if the extra logging config is using oslo_log's ContextFormatter
as formatter for some handler, the logging_context_format_string
option and the like do in fact still apply.

Instead, give the `log-date-format` as example option that does apply
when log-config-append is enabled and clarify that some options
are used with ContextFormatter only.

Change-Id: I917cdf8799ac0fb5b49873d5c7bccd957a42f347