Add logging guidelines based on previous spec
Convert the logging guidelines from the previous cross-project spec:

  https://specs.openstack.org/openstack/openstack-specs/specs/log-guidelines.html

so that they read as documentation rather than a spec, in order to
encourage developers down the right lines.

Also port a section on variable interpolation from oslo.i18n.  It was
misplaced there because it wasn't specific to i18n, and anyway logs
are no longer translated so it had to be removed from there by
Ib86013ff5e7b.

Change-Id: I3b8299b7a53e9e22c507930aa7a37c81ebcd580e