Revert "Remove the usage of the Eventlet debug feature from oslo.log."
This reverts commit 8705f67bfbd733edf923003c0a5850e7a4a5735b.

Reason for revert: Reverting this change as it is causing failures across multiple projects (nova and designate at least). OpenSearch is showing 580 related failures in the last 24 hours. As the comment states, this was an intentional setting based on how this "mutex" is implemented. I think this needs to stay in the code until we can move oslo service off of eventlet.

Related-Bug: #2072627

Change-Id: I562dc23c6ca41dcef6f40127f071b8422f677a0f